backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
boto3==1.35.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
openai>=1.12.0
httpx>=0.27.0
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.2

